+++ links-tests/Test/Xrefcheck/FtpLinks.hs
@@ -15,7 +15,7 @@ import Test.Tasty (TestTree, askOption, testGroup)
 import Test.Tasty.HUnit (assertBool, assertFailure, testCase, (@?=))
 import Test.Tasty.Options as Tasty (IsOption (..), OptionDescription (Option), safeRead)
 
-import Xrefcheck.Config (Config (cVerification), VerifyConfig (vcCheckLocalhost), defConfig)
+import Xrefcheck.Config (Config (cVerification), VerifyConfig (vcIgnoreRefs), defConfig)
 import Xrefcheck.Core (Flavor (GitHub))
 import Xrefcheck.Verify
   (VerifyError (..), VerifyResult (VerifyResult), checkExternalResource, verifyErrors)
@@ -42,7 +42,7 @@ instance IsOption FtpHostOpt where
 
 
 config :: VerifyConfig
-config = (cVerification $ defConfig GitHub) { vcCheckLocalhost = True }
+config = (cVerification $ defConfig GitHub) { vcIgnoreRefs = [] }
 
 test_FtpLinks :: TestTree
 test_FtpLinks = askOption $ \(FtpHostOpt host) -> do